feat: 使用事务处理帖子点赞和通知逻辑

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -59,22 +59,49 @@ export async function POST(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    await prisma.like.upsert({
-      where: {
-        // 使用schema.prisma 中 Like 模型设置的@@unique([userId, postId])复合唯一约束 自动生成的名称userId_postId进行查询
-        userId_postId: {
+    // 根据当前帖子id查询帖子用户id
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
+      },
+    });
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    // 多个数据库操作组合成一个事务。要么全部成功或全部失败
+    await prisma.$transaction([
+      prisma.like.upsert({
+        where: {
+          // 使用schema.prisma 中 Like 模型设置的@@unique([userId, postId])复合唯一约束 自动生成的名称userId_postId进行查询
+          userId_postId: {
+            userId: loggedInUser.id,
+            postId,
+          },
+        },
+        // 如果用户之前没有点赞过这个帖子，就创建一个新的点赞记录。
+        create: {
           userId: loggedInUser.id,
           postId,
         },
-      },
-      // 如果用户之前没有点赞过这个帖子，就创建一个新的点赞记录。
-      create: {
-        userId: loggedInUser.id,
-        postId,
-      },
-      // 如果用户之前已经点赞过这个帖子，就不做任何改变（因为 update 为空）。
-      update: {},
-    });
+        // 如果用户之前已经点赞过这个帖子，就不做任何改变（因为 update 为空）。
+        update: {},
+      }),
+      ...(loggedInUser.id !== post.userId // 点赞的用户不是帖子作者则创建一个通知
+        ? [
+            prisma.notification.create({
+              data: {
+                issuerId: loggedInUser.id,
+                recipientId: post.userId,
+                postId,
+                type: "LIKE",
+              },
+            }),
+          ]
+        : []),
+    ]);
 
     return new Response();
   } catch (error) {
@@ -94,12 +121,34 @@ export async function DELETE(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    await prisma.like.deleteMany({
-      where: {
-        userId: loggedInUser.id,
-        postId,
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
       },
     });
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    // 删除用户的点赞记录和与该点赞相关的通知
+    await prisma.$transaction([
+      prisma.like.deleteMany({
+        where: {
+          userId: loggedInUser.id,
+          postId,
+        },
+      }),
+      prisma.notification.deleteMany({
+        where: {
+          issuerId: loggedInUser.id,
+          recipientId: post.userId,
+          postId,
+          type: "LIKE",
+        },
+      }),
+    ]);
 
     return new Response();
   } catch (error) {
